EXM-40615: activate "Restart merge" even for empty root.
Details: when we used "Resolved using mine" and there's no
staged/unstaged file, the "Restart merge" action should be available in
the contextual menu shown for the childless root. Also removed the
message that tells you that all conflicts are resolved when you restart
the merging. Moreover, fixed a tree rendering problem when a resource
ceases to exist after restarting the merge.